Better temporary fix to the add 'MN' bug in script -- split links into separate lines.
Temporary fix: start a new line between any references that are in the
same paragraph. Be careful in footnote, convert script screws up the
marginnote if there is a line break.

Still a problem in CMA03-4 near "proof employs directly" (I added MN to
regular file to fix this case)